+++ src/core/msg.js
@@ -57,34 +57,3 @@ export function registerMsg(fulltypename, messagetype) {
     messagetype.typename = fulltypename
     typeregistry.set(fulltypename, messagetype.prototype)
 }
-
-export class RegistrationRequest {
-    constructor(actoraddr) {
-        this.actoraddr = actoraddr
-    }
-}
-registerMsg("CircoCore.RegistrationRequest", RegistrationRequest)
-
-export class Registered {
-    constructor(actoraddr, accepted) {
-        this.actoraddr = actoraddr
-        this.accepted = accepted
-    }
-}
-registerMsg("CircoCore.Registered", Registered)
-
-export class NameQuery {
-    constructor(name) {
-        this.name = name
-    }
-}
-registerMsg("CircoCore.NameQuery", NameQuery)
-
-export class NameResponse {
-    constructor(query, handler) {
-        this.query = query
-        this.handler = handler
-    }
-}
-registerMsg("CircoCore.NameResponse", NameResponse)
-
